+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ Version 0.4.3
 
 *yet to be released*
 
+- More performance improvements to ``singlefile``-storage.
+
 Version 0.4.2
 =============